refactored more, fixed some bugs, started work on tabular
