+++ openml_pytorch/callbacks.py
@@ -4,6 +4,7 @@ import re
 from typing import Iterable
 
 from matplotlib import pyplot as plt
+import numpy as np
 import torch
 
 _camel_re1 = re.compile("(.)([A-Z][a-z]+)")
@@ -66,8 +67,7 @@ def combine_scheds(pcts, scheds):
 
 # def cos_1cycle_anneal(start, high, end):
 #     return [sched_cos(start, high), sched_cos(high, end)]
-def accuracy(out, yb):
-    return (torch.argmax(out, dim=1) == yb.long()).float().mean()
+
 
 
 class Callback:
@@ -140,3 +140,86 @@ class Recorder(Callback):
 def camel2snake(name):
     s1 = re.sub(_camel_re1, r"\1_\2", name)
     return re.sub(_camel_re2, r"\1_\2", s1).lower()
+
+
+class TrainEvalCallback(Callback):
+    def begin_fit(self):
+        self.run.n_epochs = 0
+        self.run.n_iter = 0
+
+    def after_batch(self):
+        if not self.in_train:
+            return
+        self.run.n_epochs += 1.0 / self.iters
+        self.run.n_iter += 1
+
+    def begin_epoch(self):
+        self.run.n_epochs = self.epoch
+        self.model.train()
+        self.run.in_train = True
+
+    def begin_validate(self):
+        self.model.eval()
+        self.run.in_train = False
+
+
+class CancelTrainException(Exception):
+    pass
+
+
+class CancelEpochException(Exception):
+    pass
+
+
+class CancelBatchException(Exception):
+    pass
+
+
+class AvgStats:
+    def __init__(self, metrics, in_train):
+        self.metrics, self.in_train = listify(metrics), in_train
+
+    def reset(self):
+        self.tot_loss, self.count = 0.0, 0
+        self.tot_mets = [0.0] * len(self.metrics)
+
+    @property
+    def all_stats(self):
+        return [self.tot_loss.item()] + self.tot_mets
+
+    @property
+    def avg_stats(self):
+        return [o / self.count for o in self.all_stats]
+
+    def __repr__(self):
+        if not self.count:
+            return ""
+        return f"{'train' if self.in_train else 'valid'}: {self.avg_stats}"
+
+    def accumulate(self, run):
+        bn = run.xb.shape[0]
+        self.tot_loss += run.loss * bn
+        self.count += bn
+        for i, m in enumerate(self.metrics):
+            self.tot_mets[i] += m(run.pred, run.yb) * bn
+
+
+class AvgStatsCallBack(Callback):
+    def __init__(self, metrics):
+        self.train_stats, self.valid_stats = AvgStats(metrics, True), AvgStats(
+            metrics, False
+        )
+
+    def begin_epoch(self):
+        self.train_stats.reset()
+        self.valid_stats.reset()
+
+    def after_loss(self):
+        stats = self.train_stats if self.in_train else self.valid_stats
+        with torch.no_grad():
+            stats.accumulate(self.run)
+
+    def after_epoch(self):
+        print(self.train_stats)
+        print(self.valid_stats)
+
